add get station by company and gov endpoint to passenger side to be displayed in menufield into passenger mobile interfaces

+++ DTOs/Station/CreateStationDto.cs
@@ -15,7 +15,7 @@ namespace Darb.Api.DTOs.Station
         public TimeSpan DurationToEndStation { get; set; }
 
         [Required]
-        public double ExtraFee { get; set; }
+        public decimal ExtraFee { get; set; }
 
         public int CityId { get; set; }
         public int GovernorateId { get; set; }
